KAN-22 refactor: lift folder to top level
