+++ src/shared/components/dropDownMenu/index.ts
@@ -2,5 +2,5 @@ export { DropDownMenu } from './dropDawnMenu'
 export type { DropDownMenuProps } from './dropDawnMenu'
 export * as DropdownMenu from '@radix-ui/react-dropdown-menu'
 export { DropDownMenuItem } from './dropDownMenuItem/DropDownMenuItem'
-export { DropDownMenuLabel } from './dropDownMenuItem/dropDownMenuLabel'
+export { DropDownMenuLabel } from './dropDownMenuLabel'
 export { DropDownSeparator } from './dropDownSeparator'
